refactor: ReuqestHandler 책임 분리
