+++ app/src/main/java/nextstep/jwp/WebServer.java
@@ -1,5 +1,7 @@
 package nextstep.jwp;
 
+import nextstep.jwp.server.ControllerDispatcher;
+import nextstep.jwp.server.RequestHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -34,7 +36,7 @@ public class WebServer {
     private void handle(ServerSocket serverSocket) throws IOException {
         Socket connection;
         while ((connection = serverSocket.accept()) != null) {
-            new Thread(new RequestHandler(connection)).start();
+            new Thread(new RequestHandler(connection, ControllerDispatcher.getInstance())).start();
         }
     }
